SimpleTasker + Refactor SimpleProfiler
SimpleProfiler is now more threadsafe, I don't think is completely although it was made only for single threaded architecture (open and close scopes should be done in singe functions and hierarchichally)

As it stands it justcloses last open scope so it could be tampered with easily if not closed properly.
Ideal way would be to create variables that destruct, but don't think is a clean solution either. Probably better to use external profiling tools such as Optick or that one made with imgui I don't rememeber.

Macro1::{ var scope

Macro2::~scope}