Shell Executales for Linux Builds
For now I will try to create shell scripts that will perform the build commands for the target platforms

This is because it gets tedious to go into terminal and do that for evey build.

On android I hav eto perform extra commands, such as push and run for testing (which fail if there is no device so no problem really. Probably more commands i still don't know I will have to use more commands to opend and prepend makefile from android glue sooooo it should be way more useful in the long run.