Added Android Glue
Trying compilation of the SO files based on procesor architecture.
A lot has been learned on how to create secondary targets and be compiled.
Current issue is linking to android, EGL and log library. Technically there is a flag that adds directory of the version to the search path for this linking process (-L....ANDROIDVERSION) but still it does not find any, even if hard writting libandroid.so ,etc

I need to do more breaks and refresh better because some things are found when looking with different approaches. Test Work 2h max and take at least 1h breaks per 2h. Learn to reset better.