Install issue solved
The issue was that when compressing I was compressing the whole folder instead of contents, as AndroidManifest.xml has to be in root it was generating errors.

Currently there is an issue with linking to libc++_shader.so, which I've seen somewhere how to properly add it into the project but not so sure, for tomorrow.

Remember to adb logcat before starting and follow fatal messages (crash).
Will have to add support for android logs later on.
Probably more libraries and code fail but will get to it step by step.