CMAKE more complete
Added the other targets to generate the apk, manifests, push, uninstall and cleanup

C++17 libraries are not fully or really not implemented, not even experimental. For that it is probable that more steps are required to use the llvm-libc++ experimental or something, but including directly generates errors and linking against static did not work either. -For now no fs for android

Issue on run remains the same. Seems like the library si not well included, no tlinked properly but issue remains even when adding app_dummy(), which throws another error.

Next test would be to also include android directories to glue or something. Seems like c file is discarded, maybe we should set compiler language to C instead of CXX for compiling that