Small Naming Changes
From what I have read on the SDL wiki, I don't like how I have to setup SDL to work with the project. Mostly it tries to setup the whole project to build with SDLs systems which I do not favor for my own flexibility.

My ideal way would be to hold the source and headers for inclusion in VSCode (Intellisense) then build th erespective binaries for each platform according to the actual parts of SDL used (custom tooclhain with just the necessary dlls?)