Added WinToWin Build
issue with library was that I had different naming for functions than declaration, making issues when calling them. Will probably work fine now in Linux too.

Have to properly prepare the cmake module for windows to windows (and later windows to others), as it seems it will be tricky.

Linking to speciffic libraries will be another thing to take into consideration properly as it seems to be a bit ugly both in windows and linux (user provided/non-native libraries).