Reordering Project and pulling and building only needed of GLFW
Now GLFW can be fetched with the cmake building and will only build the dll required.
The issue now is where it will be built on different platforms and if I will have to generate multiple folderpaths

For now, the build will now be in ./builds/_dep/lib-build/src/CMAKE_BUILD_TYPE
But that is only tested on windows (will have to check on Linux and adapt, maybe on MacOS some day).
CMAKE_BUILD_TYPE is not cached for build on windows, the binaries are built directly for Debug, MinSizeRelease, Release and something else. You have to specify on --build which one to, makes it a bit hard to create proper build dynamically at first glance.