Building again for Linux and Android
After doing the reorganization of the cmake modules (which is getting convoluted) I made some name changes and errors.

Also, now build for android is not working as before, probably issue with linking to some things or not doing things properly with library inclusion.
The error is based in the linked not finding the function so I believe linking to FlyLib is not being done correctly from TempEngine on Android

-fvisibility=hidden was making functions not available to the linked things, I don't know why because I don't understand the flag properly