Windows Window base working
Display now initiates and closes window for GLFW, I don't like how i have to predefine GLFWwindow for the callback but maybe I should not expose single window 

I believe, the right practice is to initialize each part when needed isntead of calling a whole init for library, maybe leave it as a default option but better a way to send the initialization options.

Learn Flag management / setup for libraries and things and use it to do it on the library.

Make the glfw_windows be a workflow for every platform (hold all windows). Each window should be another struct that holds the size, title and something else needed (flags?)...