Tweaks for code correctnes and working on linux/android
Actually there were a couple of issues with memory management in android, be more careful when moving code around and finally I got bit in the arse for not properly checking what happens if logsize was exceeded. Now they are not process, have to do something in the future.

For no wthis is okay, make something for back, resume,....

Currently it is an infinite loop that forces a reset on the device because it is using to much or something.