ImGui basic implementation into engine
Now imgui comes with engine rather than library
Library has access to everything it implements publicly, that way when you require some intricacies to implement things, you can do it with what it allows to implement.

In this case, ImGui requires specifically to load the OpenGL functions on the main process, else it does not want to work. Ej: call gladLoadGL and Init_ImplGlfw_InitForOpenGL that requires main GLFWwindow

Now to mak eit work on linux