Functions inside structs/classes have to be declared to export
I was having issues with FLY_Timer, Adding FL_API was the solution, as ALL functions regardless of where they are declared have to be declared for export/import in any case.

On linux and android it was working because the compiler might already export the whole struct not just data.

GLFW now build into the desired folder, instead of on build, I have to learn how to specify not to create build and sub build folder or use ExternalProjectAdd...