FlyRender Base
Using LearnOpenGL as base and will throw my own types, structs and functions to make it mroe complete along side it.

Currently he simplest triangle pipeline is done, but I can't get to render the triangle, I don't know really why. Probably something I am missing, but OpenGL throws no errors and no crashes. Will try to make sure the vertices are not wrong or something

Compiling GLAD as static is the better way to get it to work for desktop apparently and debug, can be shared on release, not on debug I don't know why.