More Program Control
Better management of buffers (minimally) through setting the buffer sizes.
Now Drawing with a Program requires to pass the buffer you will draw. When passing a buffer that does nto match the program, OpenGL will not execute it or throw graphic errors because data mismatch, but now the attributes are mroe dynamic (but names for vertex attributes are hardcoded, don't know about that).