MathgeoLib with FlyLib
Now you can call MathgeoLib which will be the main math library for flylib (it makes not sense that I make my own based on first lack of knowledge and then time)

Deleted my shitty types (int2,3,... and typedefs for sized ints) and using more standard types.

MathGeoLib is compiled and added to the project easily now, I should try to add groups for the sources properly, headers are now part of flylib and can be seen better (although they were public, they were not accessible from within the explorer).

Stopped printf, because it makes a mess on console and I want to not open a console if possible, use the debug options per system, will help with debugging with gdb (will ttry to add gdb support for android).

Now, I should focus on the main cases for filesystem and do the build helper for flylib.

After that, add framebuffers and we should have the first most basic release of flylib:
Build Helper, the base library
This way it is possible to build upon it, use the types and actually do things in plain c++.