Asset Packing is hardcoded currently
Installation should be separate from Compilation for Android (duh), so that we separate the copy process properly and then package the app properly.

For now I will use the folders as provided and ge tit working.

Asset loading works for 1 instance, next time it tries to acces my_app apparently it crashes, which is not very nice.

If I load 1 asset it will crash on makeContextMain, else it crashes on next asset load.