Putting Load Implementation in SqueLib
Tried to get SOIL2 to build with SqueLib but did not go as expected, It 100% requires you to build the static lib and then link the version accordingly instead of being able to build it by yourself with your toolchain.

In a future where I need more texture formats or specialized (dds), I will have to make the full tooclhain for SOIL in cmake, for now I will only use png and jpgs with stbimage, should be more than enough.