Personal and Engine Tests
This example will be used for me to do the performance I think are necessary or interesting to do and then apply properly.

Today I found out that reserve is shit compared to resize and setting specific memory. Literally like 2x performance and same actual memory usage somehow.

Direct Push_Back is slow even with a previous reserve of required size.
If for resize is much faster than branching resize.