Copy Constructor for sque_vec
Because of no properly made copy constructor, data was no tproperly copied, instead an array would point to the new array now and would not properl yhave the data realigned -> That was super important for vectors of vectors which I was using (it is not good but it works).

Now does not crash on multi data, probably because data was getting corrupted most probably.