LearnOpenGL Shaders/Uniforms/Attributes
Found out that I had not binded the attribute_object on draw_vertices.

HAve to generate a couple defaults for generating things because when doing small things it gets too verbose I feel.

I don't want someone to have to write all the setup with attributes for something that does nto need to generate a lot of attributes.

A simple triangle and should not take so many functions, it is not that much and is pretty self explanatory each step but the process is not fun i guess.