Android stbi fails to load PNG files
There is not a PNG error, but a JPEG error, and file seems to be loading more or less. The issue is that when it is returned the pointer points to "\0" and does not hold actual info for OPENGL to parse it properly.

Because how I debugon android, I can't really get proper info on how that happens and what the issue is (maybe if there is an error recorder it return value null) but the biggest issue is that is a silent error.