Camera Movement android
ImGui Android interaction was a bit fucked. The issue coming from pointer0 not sending basic info about press and such.

There is an issue with the bypass is that it breaks gestures and pointers... or if you want to have them at the same time it will be hard to control...