GLSL Ver is global
There are not clean enough but will work because it is an undergrad project. It will not be production ready but it still makes sense to go all the way.

MultiWindow and MultiViewport have to be forgone, it is too much complication and typical use case even in android will be 1 window.

Next step is to clean and generate documentation of library.