Added SoLoud to take care of audio
SoLoud is included directly in source project it is not the greatest option but will work for now. Better solution would be to make a static version to compile and link to source project.

Works for basic audio, will now test on linux and android and call it quits for audio and go to make GLFW and ImGui local repos instead of Fetched because it makes no sense (more memory and more time to download).

Documentation was fucked because I missed a .. instead of . for including markdeep...