Quick Sort done
I struggled a lot to get it working and understand it a bit.
The core concept is very simple but the actual implementation without generating a ton of memory with temp arrays is trickier than I thought.

It pains me that I am clearly not a proper C++ developer with no mathematical and CS approach