sque_vector does not work well with arrays
arrays don't have default class destructor as a type, thus will not e able to be destructed...

In any case, tested for memory leaks with valgrind and there was a couple bytes lost because of not cleaned things but most issues pointed come from glfw, imgui,... that must have their own mallocs or a way to deal with allocations that valgrind gets as leaks.

In any case, a otn of todos and have to do a big search on what is causing that memory leak over long periods of time.