More data structures
Separated vector from a freeable vector (that pushes where space is available like a heap. Also a double linked list.

A step is to delete the hardcoded arrays I had for the attributes. I think Uniforms are missing to be deleted and put inside the structure.

Right now each struct that requires attributes will have a freeable vector for this, it makes no sense to worry about memory compactness here on initialization while losing practicality of access.