Moving some Extrenal to Extra and rewrite
SoLoud and GLMath ought to be extra libraries as they aren't part of squelib but nice additions to compile with.

Had to specifcy the binary folder for GLM as it was out of tree.
SoLoud should also be compiled as a static library probably.

Have to update examples, add ImGui base to Extra and then finish the basic render pipeline for SqueMaker, including the shader thing to read inputs and transform into nodes. Then Add proper write and read on android for Disk, Tinygltf, loader into own format, then Drawable.

With that TFG should be done.