Info about Filewatch news
When taking a look at files, it will return a vector of initialized char*[512] which is excessive in size but the issue furthers that it is initialized memory that the user should delete, besides making a class that deals with it and thus making it less generic i don;'t see a solution to that. Should probably be a list instead of a vec because they are string of high size, you will have a ton of cache misses either way probably.